# ENH Optimize runtime for IsolationForest