--- a/sklearn/ensemble/_iforest.py
+++ b/sklearn/ensemble/_iforest.py
@@ -309,7 +309,7 @@ class IsolationForest(OutlierMixin, BaseBagging):
             max_samples,
             max_depth=max_depth,
             sample_weight=sample_weight,
-            check_input=True,
+            check_input=False,
         )
 
         if self.contamination == "auto":