<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/scikit-learn/scikit-learn/blob/main/CONTRIBUTING.md
-->

#### Reference Issues/PRs

Towards #19275

#### What does this implement/fix? Explain your changes.
As shown [here](https://github.com/scikit-learn/scikit-learn/issues/19275#issuecomment-811615662) in the original issue discussion, the `check_input` argument is set to False for `Forest` classes, while for bagging estimators it's left as default value `True`, therefore we're validating the data repeatedly. 

The proposed fix adds a `check_input`  argument to the `ensemble._bagging._parallel_build_estimators`, which can be set to False when the base estimator actually supports that argument during the fit process.

#### Performance Impact
Code used for profiling:
```python
from sklearn.datasets import make_classification
from scipy.sparse import csc_matrix, csr_matrix
from sklearn.ensemble import IsolationForest

X, y = make_classification(n_samples=50000, n_features=1000)
X = csc_matrix(X)
X.sort_indices()
IsolationForest(n_estimators=10, max_samples=256, n_jobs=1).fit(X)
```

Before (total time: 6.78s)
![image](https://user-images.githubusercontent.com/16646940/163718907-97e636c6-0b06-497d-9143-84cfd3f83cc2.png)

After (total time: 5.017s)
![image](https://user-images.githubusercontent.com/16646940/163719002-7f362616-e9db-4c20-942b-a621ba347b7f.png)



<!--
Please be aware that we are a loose team of volunteers so patience is
necessary; assistance handling other issues is very welcome. We value
all user contributions, no matter how minor they are. If we are slow to
review, either the pull request needs some benchmarking, tinkering,
convincing, etc. or more likely the reviewers are simply busy. In either
case, we ask for your understanding during the review process.
For more information, see our FAQ on this topic:
http://scikit-learn.org/dev/faq.html#why-is-my-pull-request-not-getting-any-attention.

Thanks for contributing!
-->