@MaxwellLZH 2022-04-18
The benchmark result for dense data shows smaller improvement compared to sparse inputs

**Before (8.29s):**
![image](https://user-images.githubusercontent.com/16646940/163738947-33587866-6023-40ba-8936-d5e14610d4db.png)


**After (7.92s):**
![image](https://user-images.githubusercontent.com/16646940/163738869-e14081c1-d4c5-4d7f-8714-61621af186e9.png)