@MaxwellLZH 2022-04-23
After more profiling, it seems like the indexing operation of `estimator.fit(X[:, features, y)` in `_bagging.py::_parallel_build_estimators` is quite expensive, I tried removing the indexing operation when `max_features` is equal to feature numbers using the following logic:
```python
# only index arrays when needed
X_ = X if max_features == X.shape[1] else X[:, features]
estimator_fit(X_, y, sample_weight=curr_sample_weight)
```
and see improvements for both sparse and dense inputs. 

While the problem will still exist when `max_features` is not set to 1.0 for `IsolationForest`. I'm wondering since we're randomly pick a single feature to split anyway, can we always set the `max_features` to be 1.0 for `IsolationForest` so we can get a faster run time.


**Sparse Input:**
Before (4.97s)
![image](https://user-images.githubusercontent.com/16646940/164909922-70a89054-a81e-4931-acca-46e4779d8eba.png)


After (0.241s)
![image](https://user-images.githubusercontent.com/16646940/164909946-4c070410-0176-43ed-a94b-cd9e08afb1af.png)



**Dense Input:**
Before (7.99s)
![image](https://user-images.githubusercontent.com/16646940/164909872-3cec6573-efd9-4baf-83a2-d47caaec7e14.png)

After (5.19s)
![image](https://user-images.githubusercontent.com/16646940/164909792-52b711b6-0280-44cc-8566-036ba0043162.png)