@ogrisel 2022-04-29
Merging this one. The real fix for the sparse case should happen in a follow-up PR probably targeting 1.2.